Add the initial version of the Aprimo app.
This is the initial version of the Aprimo App for Contentful. It supports launching the [Aprimo Content Selector](https://developers.aprimo.com/digital-asset-management/aprimo-integration-tools/aprimo-content-selector/) from a field entry in Contentful and then storing the information about the selected asset from Aprimo DAM within Contentful.

The Aprimo Content Selector is launched in a popup window due to technical limitations of embedding within a Contentful dialog. However there is code to use a Contentful dialog in here that we can enable once embedding is possible with Aprimo Content Selector.

The structure of the data from Aprimo DAM that is stored in Contentful is:

```typescript
interface AprimoAsset {
  id: string;
  title: string;
  rendition: AprimoRendition;
}

interface AprimoRendition {
  id: string;
  publicuri: string;
}
```

This is built on v2.0.0 of [@contentful/dam-app-base](https://www.npmjs.com/package/@contentful/dam-app-base).